Use type float32 instead of float64 for regularization weights.

+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/regularization.hpp
@@ -21,7 +21,7 @@ namespace boosting {
              *
              * @return The weight of the regularization term
              */
-            virtual float64 getWeight() const = 0;
+            virtual float32 getWeight() const = 0;
     };
 
 }
